+++ code/app/src/main/java/com/example/project_seekdeep/Comments/CommentAdapter.java
@@ -1,6 +1,5 @@
-package com.example.project_seekdeep;
+package com.example.project_seekdeep.Comments;
 
-import android.graphics.drawable.Drawable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,6 +8,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.project_seekdeep.R;
 
 import java.util.ArrayList;
 import java.util.List;
