+++ code/app/src/main/java/com/example/project_seekdeep/Followings/FollowRequest.java
@@ -1,4 +1,4 @@
-package com.example.project_seekdeep;
+package com.example.project_seekdeep.Followings;
 
 import com.google.firebase.firestore.DocumentReference;
