+++ code/app/src/main/java/com/example/project_seekdeep/Helpers/NotificationHandler.java
@@ -1,4 +1,4 @@
-package com.example.project_seekdeep;
+package com.example.project_seekdeep.Helpers;
 
 import android.content.Context;
 import android.util.Log;
