+++ code/app/src/main/java/com/example/project_seekdeep/Helpers/SocialSituations.java
@@ -1,4 +1,4 @@
-package com.example.project_seekdeep;
+package com.example.project_seekdeep.Helpers;
 
 import androidx.annotation.NonNull;
