+++ code/app/src/main/java/com/example/project_seekdeep/Moods/MoodFiltering.java
@@ -1,10 +1,10 @@
-package com.example.project_seekdeep;
+package com.example.project_seekdeep.Moods;
 
-import java.lang.reflect.Array;
+
+import com.example.project_seekdeep.Helpers.EmotionalStates;
+
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
