+++ code/app/src/main/java/com/example/project_seekdeep/Moods/MoodProvider.java
@@ -1,6 +1,4 @@
-package com.example.project_seekdeep;
-
-import android.graphics.Movie;
+package com.example.project_seekdeep.Moods;
 
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
